ggc-lambda-function/gglcsecret-func/requirements.txt
requests
greengrasssdk
mysql-connector-python